Fix skip statement for error handler test
We do not require a specific version of PHP, we only require PHPUnit 5.1. Previously, the test might have been run with PHP 7.0 and PHPUnit 4.8.x and then failed instead of being skipped.

PHPUnit 5.1 is required for the test to run successfully, PHPUnit 4.8.13 is required for PHPUnit_Runner_Version::series() and is included to run tests on PHP 5.5 and 5.4.